Fix for "ModelExplorer can cannot insert InstantRegisterClasses on an "empty" unit" under D2009 + Added logic to explicitly set the StartPos and EndPos of the newly added InitializationSection.
Note: The design of the original logic is flawed and fails to properly initialize the position of newly inserted Code Sections. The code worked under previous versions of Delphi, but apparently largely by accident. A related coding error in TInstantCodeModifier.InsertObjectText corrected in Rev 815 appears to have masked this defect.